services/auth-service/requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
prometheus-client
prometheus-fastapi-instrumentator

